+++ functions/GetStuList.fnc.php
@@ -187,7 +187,15 @@ function GetStuList( &$extra = array() )
 				p.FIRST_NAME,p.LAST_NAME,p.MIDDLE_NAME,
 				sjp.STUDENT_RELATION,sjp.EMERGENCY,sjp.CUSTODY,pjc.TITLE,pjc.VALUE,a.PHONE,sjp.ADDRESS_ID ';
 
-			$extra2['FROM'] = ',ADDRESS a,STUDENTS_JOIN_ADDRESS sja LEFT OUTER JOIN STUDENTS_JOIN_PEOPLE sjp ON (sja.STUDENT_ID=sjp.STUDENT_ID AND sja.ADDRESS_ID=sjp.ADDRESS_ID)
+			$extra2['FROM'] = '';
+
+			if ( empty( $extra['addr'] ) )
+			{
+				// Fix SQL error table ADDRESS specified more than once.
+				$extra2['FROM'] .= ',ADDRESS a';
+			}
+
+			$extra2['FROM'] .= ',STUDENTS_JOIN_ADDRESS sja LEFT OUTER JOIN STUDENTS_JOIN_PEOPLE sjp ON (sja.STUDENT_ID=sjp.STUDENT_ID AND sja.ADDRESS_ID=sjp.ADDRESS_ID)
 				LEFT OUTER JOIN PEOPLE p ON (p.PERSON_ID=sjp.PERSON_ID)
 				LEFT OUTER JOIN PEOPLE_JOIN_CONTACTS pjc ON (pjc.PERSON_ID=p.PERSON_ID) ';
 
@@ -239,8 +247,9 @@ function GetStuList( &$extra = array() )
 				$extra2['SELECT_ONLY'] = 'ssm.STUDENT_ID,p.PERSON_ID,p.FIRST_NAME,p.LAST_NAME,p.MIDDLE_NAME,
 					sjp.STUDENT_RELATION,sjp.EMERGENCY,sjp.CUSTODY,pjc.TITLE,pjc.VALUE,a.PHONE,sjp.ADDRESS_ID ';
 
-				if ( empty( $extra2['FROM'] )
-					|| stripos( $extra2['FROM'], 'ADDRESS a' ) === false )
+				if ( ( empty( $extra2['FROM'] )
+						|| stripos( $extra2['FROM'], 'ADDRESS a' ) === false )
+					&& empty( $extra['addr'] ) )
 				{
 					// Fix SQL error table ADDRESS specified more than once.
 					$extra2['FROM'] .= ',ADDRESS a';
@@ -1269,8 +1278,7 @@ function appendSQL( $sql, $extra = array() )
 	}
 
 	// Address (City, State, Zip code) (contains, case insensitive).
-	if ( isset( $_REQUEST['addr'] )
-		&& $_REQUEST['addr'] !== '' )
+	if ( ! empty( $_REQUEST['addr'] ) )
 	{
 		$sql .= " AND (LOWER(a.ADDRESS) LIKE '%" . mb_strtolower( $_REQUEST['addr'] ) .
 			"%' OR LOWER(a.CITY) LIKE '" . mb_strtolower( $_REQUEST['addr'] ) .